+++ backend/fns/src/main/java/ssafy/fns/domain/member/service/dto/MemberResponseDto.java
@@ -13,7 +13,7 @@ import ssafy.fns.domain.member.controller.dto.MemberProfileRequestDto;
 public class MemberResponseDto {
 
     private String nickname;
-    private Integer age;
+    private Long age;
     private Long height;
     private Long weight;
     private String gender;
